tp4_sd/trabalho_sd/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-bootstrap-datepicker-plus==3.0.5
psycopg2-binary==2.9.1
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
